refactor system (remove derivatives, make parametric)
