+++ src/quantum_systems/quantum_systems.jl
@@ -1,68 +1,6 @@
-module QuantumSystems
-
-export AbstractQuantumSystem
 export QuantumSystem
 export OpenQuantumSystem
 export VariationalQuantumSystem
-export TimeDependentQuantumSystem
-
-export get_drift
-export get_drives
-
-using ..Isomorphisms
-using ..QuantumObjectUtils
-
-using LinearAlgebra
-using SparseArrays
-using TestItems
-using ForwardDiff
-
-# TODO: 
-# 1. Notice that a -> [] and a, t -> [zeros(size(G_drift))]?
-# 2. What to do about if else function definitions? Bad practice at parse time.
-
-
-function generator_jacobian(G::Function)
-    return function ∂G(a::AbstractVector{Float64})
-        ∂G⃗ = ForwardDiff.jacobian(a_ -> vec(G(a_)), a)
-        dim = Int(sqrt(size(∂G⃗, 1)))
-        return [reshape(∂G⃗ⱼ, dim, dim) for ∂G⃗ⱼ ∈ eachcol(∂G⃗)]
-    end
-end
-
-# ----------------------------------------------------------------------------- #
-# AbstractQuantumSystem
-# ----------------------------------------------------------------------------- #
-
-"""
-    AbstractQuantumSystem
-
-Abstract type for defining systems.
-"""
-abstract type AbstractQuantumSystem end
-
-# ----------------------------------------------------------------------------- #
-# AbstractQuantumSystem methods
-# ----------------------------------------------------------------------------- #
-
-"""
-    get_drift(sys::AbstractQuantumSystem)
-
-Returns the drift Hamiltonian of the system.
-"""
-get_drift(sys::AbstractQuantumSystem) = sys.H(zeros(sys.n_drives))
-
-"""
-    get_drives(sys::AbstractQuantumSystem)
-
-Returns the drive Hamiltonians of the system.
-"""
-function get_drives(sys::AbstractQuantumSystem)
-    H_drift = get_drift(sys)
-    # Basis vectors for controls will extract drive operators
-    return [sys.H(I[1:sys.n_drives, i]) - H_drift for i ∈ 1:sys.n_drives]
-end
-
 
 # ----------------------------------------------------------------------------- #
 # QuantumSystem
@@ -76,7 +14,6 @@ A struct for storing quantum dynamics and the appropriate gradients.
 # Fields
 - `H::Function`: The Hamiltonian function, excluding dissipation: a -> H(a).
 - `G::Function`: The isomorphic generator function, including dissipation, a -> G(a).
-- `∂G::Function`: The generator jacobian function, a -> ∂G(a).
 - `levels::Int`: The number of levels in the system.
 - `n_drives::Int`: The number of drives in the system.
 
@@ -87,10 +24,9 @@ A struct for storing quantum dynamics and the appropriate gradients.
 - QuantumSystem(H::Function, n_drives::Int; kwargs...)
 
 """
-struct QuantumSystem <: AbstractQuantumSystem
-    H::Function
-    G::Function
-    ∂G::Function
+struct QuantumSystem{F1<:Function, F2<:Function} <: AbstractQuantumSystem
+    H::F1
+    G::F2
     n_drives::Int
     levels::Int
     params::Dict{Symbol, Any}
@@ -121,17 +57,14 @@ struct QuantumSystem <: AbstractQuantumSystem
         if n_drives == 0
             H = a -> H_drift
             G = a -> G_drift
-            ∂G = a -> 0
         else
             H = a -> H_drift + sum(a .* H_drives)
             G = a -> G_drift + sum(a .* G_drives)
-            ∂G = a -> G_drives
         end
 
-        return new(
+        return new{typeof(H), typeof(G)}(
             H,
             G,
-            ∂G,
             n_drives,
             levels,
             params
@@ -147,14 +80,13 @@ struct QuantumSystem <: AbstractQuantumSystem
         QuantumSystem(H_drift, Matrix{ℂ}[]; kwargs...)
 
     function QuantumSystem(
-        H::Function, 
+        H::F, 
         n_drives::Int; 
         params::Dict{Symbol, <:Any}=Dict{Symbol, Any}()
-    )
+    ) where F <: Function
         G = a -> Isomorphisms.G(sparse(H(a)))
-        ∂G = generator_jacobian(G)
         levels = size(H(zeros(n_drives)), 1)
-        return new(H, G, ∂G, n_drives, levels, params)
+        return new{F, typeof(G)}(H, G, n_drives, levels, params)
     end
 
 end
@@ -190,10 +122,9 @@ See also [`QuantumSystem`](@ref).
 - OpenQuantumSystem(H::Function, n_drives::Int; kwargs...)
 
 """
-struct OpenQuantumSystem <: AbstractQuantumSystem
-    H::Function
-    𝒢::Function
-    ∂𝒢::Function
+struct OpenQuantumSystem{F1<:Function, F2<:Function} <: AbstractQuantumSystem
+    H::F1
+    𝒢::F2
     n_drives::Int
     levels::Int
     dissipation_operators::Vector{Matrix{ComplexF64}}
@@ -207,7 +138,8 @@ struct OpenQuantumSystem <: AbstractQuantumSystem
         kwargs...
     )
     OpenQuantumSystem(
-        H_drift::Matrix{<:Number}, H_drives::AbstractVector{Matrix{<:Number}};
+        H_drift::Matrix{<:Number}, 
+        H_drives::AbstractVector{Matrix{<:Number}};
         dissipation_operators::AbstractVector{<:AbstractMatrix{<:Number}}=Matrix{ComplexF64}[],
         kwargs...
     )
@@ -246,17 +178,14 @@ struct OpenQuantumSystem <: AbstractQuantumSystem
         if n_drives == 0
             H = a -> H_drift
             𝒢 = a -> 𝒢_drift + 𝒟
-            ∂𝒢 = a -> 0
         else
             H = a -> H_drift + sum(a .* H_drives)
             𝒢 = a -> 𝒢_drift + sum(a .* 𝒢_drives) + 𝒟
-            ∂𝒢 = a -> 𝒢_drives
         end
 
-        return new(
+        return new{typeof(H), typeof(𝒢)}(
             H,
             𝒢,
-            ∂𝒢,
             n_drives,
             levels,
             dissipation_operators,
@@ -288,14 +217,13 @@ struct OpenQuantumSystem <: AbstractQuantumSystem
         OpenQuantumSystem(H_drift, Matrix{T}[]; kwargs...)
 
     function OpenQuantumSystem(
-        H::Function, n_drives::Int;
+        H::F, n_drives::Int;
         dissipation_operators::AbstractVector{<:AbstractMatrix{ℂ}}=Matrix{ComplexF64}[],
         params::Dict{Symbol, <:Any}=Dict{Symbol, Any}()
-    ) where ℂ <: Number
+    ) where {F <: Function, ℂ <: Number}
         G = a -> Isomorphisms.G(Isomorphisms.ad_vec(sparse(H(a))))
-        ∂G = generator_jacobian(G)
         levels = size(H(zeros(n_drives)), 1)
-        return new(H, G, ∂G, n_drives, levels, dissipation_operators, params)
+        return new{F, typeof(G)}(H, G, n_drives, levels, dissipation_operators, params)
     end
 
     OpenQuantumSystem(system::QuantumSystem; kwargs...) = OpenQuantumSystem(
@@ -305,112 +233,15 @@ struct OpenQuantumSystem <: AbstractQuantumSystem
 end
 
 # ----------------------------------------------------------------------------- #
-# TimeDependentQuantumSystem
-# ----------------------------------------------------------------------------- #
-
-# TODO: Open System
-
-"""
-    TimeDependentQuantumSystem <: AbstractQuantumSystem
-
-A struct for storing time-dependent quantum dynamics and the appropriate gradients.
-
-# Additional fields
-- `H::Function`: The Hamiltonian function with time: a, t -> H(a, t).
-- `G::Function`: The isomorphic generator function with time, a, t -> G(a, t).
-- `∂G::Function`: The generator jacobian function with time, a, t -> ∂G(a, t).
-- `n_drives::Int`: The number of drives in the system.
-- `levels::Int`: The number of levels in the system.
-- `params::Dict{Symbol, Any}`: A dictionary of parameters.
-
-"""
-struct TimeDependentQuantumSystem <: AbstractQuantumSystem
-    H::Function
-    G::Function
-    ∂G::Function
-    n_drives::Int
-    levels::Int 
-    params::Dict{Symbol, Any}
-
-    function TimeDependentQuantumSystem end
-
-    function TimeDependentQuantumSystem(
-        H_drift::AbstractMatrix{<:Number},
-        H_drives::Vector{<:AbstractMatrix{<:Number}};
-        carriers::AbstractVector{<:Number}=zeros(length(H_drives)),
-        phases::AbstractVector{<:Number}=zeros(length(H_drives)),
-        params::Dict{Symbol, <:Any}=Dict{Symbol, Any}()
-    )
-        levels = size(H_drift, 1)
-        H_drift = sparse(H_drift)
-        G_drift = sparse(Isomorphisms.G(H_drift))
-
-        n_drives = length(H_drives)
-        H_drives = sparse.(H_drives)
-        G_drives = sparse.(Isomorphisms.G.(H_drives))
-
-        H = (a, t) -> H_drift + sum(a[i] * cos(ω * t + ϕ) .* H_i for (i, (ω, ϕ, H_i)) in enumerate(zip(carriers, phases, H_drives)))
-        G = (a, t) -> G_drift + sum(a[i] * cos(ω * t + ϕ) .* G_i for (i, (ω, ϕ, G_i)) in enumerate(zip(carriers, phases, G_drives)))
-
-        function ∂G(a, t)
-            # Preallocate the Jacobian
-            ∂G_ = [zeros(eltype(G_drift), size(G_drift)) for _ in 1:n_drives + 1]
-            for i = 1:n_drives
-                ∂G_[i] = cos(carriers[i] * t + phases[i]) .* G_drives[i]
-            end
-            ∂G_[end] = sum(a * -sin(ω * t + ϕ) .* G for (ω, ϕ, G) in zip(carriers, phases, G_drives))
-            return ∂G_
-        end
-
-        # save carries and phases
-        params[:carriers] = carriers
-        params[:phases] = phases
-
-        return new(
-            H,
-            G,
-            ∂G,
-            n_drives,
-            levels,
-            params
-        )
-    end
-
-    function TimeDependentQuantumSystem(H_drives::Vector{<:AbstractMatrix{ℂ}}; kwargs...) where ℂ <: Number
-        @assert !isemtpy(H_drives) "At least one drive is required."
-        return TimeDependentQuantumSystem(spzeros(ℂ, size(H_drives[1])), H_drives; kwargs...)
-    end
-
-    TimeDependentQuantumSystem(H_drift::AbstractMatrix{ℂ}; kwargs...) where ℂ <: Number = 
-        TimeDependentQuantumSystem(H_drift, Matrix{ℂ}[]; kwargs...)
-
-    function TimeDependentQuantumSystem(
-        H::Function,
-        n_drives::Int;
-        params::Dict{Symbol, <:Any}=Dict{Symbol, Any}()
-    )
-        G = a, t -> Isomorphisms.G(sparse(H(a,t)))
-        ∂G = generator_jacobian(G)
-        levels = size(H(zeroes(n_drives)), 1)
-        return new(H, G, ∂G, n_drives, levels, params)
-    end
-
-    # TODO: tests for this constructor
-
-end
-
-# ----------------------------------------------------------------------------- #
 # VariationalQuantumSystem
 # ----------------------------------------------------------------------------- #
 
 # TODO: Open quantum systems?
 
-struct VariationalQuantumSystem <: AbstractQuantumSystem
-    H::Function 
-    G::Function
-    ∂G::Function
-    G_vars::Vector{Function}
-    ∂G_vars::Vector{Function}
+struct VariationalQuantumSystem{F1<:Function, F2<:Function, F⃗3<:AbstractVector{<:Function}} <: AbstractQuantumSystem
+    H::F1 
+    G::F2
+    G_vars::F⃗3
     n_drives::Int 
     levels::Int 
     params::Dict{Symbol, Any}
@@ -438,16 +269,14 @@ struct VariationalQuantumSystem <: AbstractQuantumSystem
         if n_drives == 0
             H = a -> H_drift
             G = a -> G_drift
-            ∂G = a -> 0
-            ∂G_vars = [a -> 0 for G in G_vars]
         else
             H = a -> H_drift + sum(a .* H_drives)
             G = a -> G_drift + sum(a .* G_drives)
-            ∂G = a -> G_drives
-            ∂G_vars = [a -> [spzeros(size(G)) for G in G_drives] for G in G_vars]
         end
 
-        return new(H, G, ∂G, G_vars, ∂G_vars, n_drives, levels, params)
+        return new{typeof(H), typeof(G), typeof(G_vars)}(
+            H, G, G_vars, n_drives, levels, params
+        )
     end
 
     function VariationalQuantumSystem(
@@ -466,23 +295,20 @@ struct VariationalQuantumSystem <: AbstractQuantumSystem
     end
 
     function VariationalQuantumSystem(
-        H::Function,
-        H_vars::AbstractVector{<:Function},
+        H::F1,
+        H_vars::F⃗2,
         n_drives::Int; 
         params::Dict{Symbol, <:Any}=Dict{Symbol, Any}()
-    )
+    ) where {F1 <: Function, F⃗2 <: AbstractVector{<:Function}}
         @assert !isempty(H_vars) "At least one variational operator is required"
         G = a -> Isomorphisms.G(sparse(H(a)))
-        ∂G = generator_jacobian(G)
         G_vars = Function[a -> Isomorphisms.G(sparse(H_v(a))) for H_v in H_vars]
-        ∂G_vars = Function[generator_jacobian(G_v) for G_v in G_vars]
         levels = size(H(zeros(n_drives)), 1)
-        return new(H, G, ∂G, G_vars, ∂G_vars, n_drives, levels, params)
+        return new{F1, typeof(G), F⃗2}(H, G, G_vars, n_drives, levels, params)
     end
 end
 
-#***********************************************************************************************#
-
+# ******************************************************************************* #
 
 @testitem "System creation" begin
     H_drift = PAULIS[:Z]
@@ -493,12 +319,6 @@ end
     @test system isa QuantumSystem
     @test get_drift(system) == H_drift
     @test get_drives(system) == H_drives
-
-    # test jacobians
-    a = randn(n_drives)
-    ∂G = system.∂G(a)
-    @test length(∂G) == system.n_drives
-    @test all(∂G .≈ QuantumSystems.generator_jacobian(system.G)(a))
 
     # repeat with a bigger system
     H_drift = kron(PAULIS[:Z], PAULIS[:Z])
@@ -511,12 +331,6 @@ end
     @test system isa QuantumSystem
     @test get_drift(system) == H_drift
     @test get_drives(system) == H_drives
-
-    # test jacobians
-    a = randn(n_drives)
-    ∂G = system.∂G(a)
-    @test length(∂G) == system.n_drives
-    @test all(∂G .≈ QuantumSystems.generator_jacobian(system.G)(a))
 end
 
 @testitem "Parametric system creation" begin
@@ -557,11 +371,6 @@ end
     @test get_drift(system) == PAULIS[:Z]
     @test get_drives(system) == [PAULIS[:X]]
 
-    # test jacobians
-    compare = QuantumSystem(PAULIS[:Z], [PAULIS[:X]])
-    a = randn(system.n_drives)
-    @test system.∂G(a) == compare.∂G(a)
-
     # test three drives
     system = QuantumSystem(
         a -> a[1] * PAULIS[:X] + a[2] * PAULIS[:Y] + a[3] * PAULIS[:Z], 3
@@ -587,13 +396,6 @@ end
     # test dissipation
     𝒢_drift = Isomorphisms.G(Isomorphisms.ad_vec(H_drift))
     @test system.𝒢(zeros(system.n_drives)) != 𝒢_drift
-
-    # test jacobians (disspiation is constant)
-    a = randn(system.n_drives)
-    ∂𝒢 = system.∂𝒢(a)
-    @test length(∂𝒢) == system.n_drives
-    @test all(∂𝒢 .≈ QuantumSystems.generator_jacobian(system.𝒢)(a))
-
 end
 
 @testitem "Open system alternate constructors" begin
@@ -655,7 +457,6 @@ end
     G_X = Isomorphisms.G(PAULIS.X)
     G_Y = Isomorphisms.G(PAULIS.Y)
     G = a[1] * G_X + a[2] * G_Y
-    ∂G_vars = [zeros(size(G_X)), zeros(size(G_Y))]
     for varsys in [varsys1, varsys2]
         @assert varsys isa VariationalQuantumSystem
         @assert varsys.n_drives == 2
@@ -663,8 +464,6 @@ end
         @assert varsys.G(a) ≈ G
         @assert varsys.G_vars[1](a) ≈ G_X
         @assert varsys.G_vars[2](a) ≈ G_Y
-        @assert varsys.∂G_vars[1](a) ≈ ∂G_vars
-        @assert varsys.∂G_vars[2](a) ≈ ∂G_vars
     end
 
     # single sensitivity
@@ -677,7 +476,6 @@ end
     @assert length(varsys.G_vars) == 1
     @assert varsys.G(a) ≈ G
     @assert varsys.G_vars[1](a) ≈ G_X
-    @assert varsys.∂G_vars[1](a) ≈ ∂G_vars
 
     # functional sensitivity
     varsys = VariationalQuantumSystem(
@@ -691,27 +489,4 @@ end
     @assert varsys.G(a) ≈ G
     @assert varsys.G_vars[1](a) ≈ a[1] * G_X
     @assert varsys.G_vars[2](a) ≈ G_Y
-    @assert varsys.∂G_vars[1](a) ≈ [G_X, zeros(size(G_Y))]
-    @assert varsys.∂G_vars[2](a) ≈ ∂G_vars
-
-end
-
-@testitem "Generator jacobian types" begin
-    GX = Isomorphisms.G(PAULIS.X)
-    GY = Isomorphisms.G(PAULIS.Y)
-    GZ = Isomorphisms.G(PAULIS.Z)
-    G(a) = GX + a[1] * GY + a[2] * GZ
-    ∂G = QuantumSystems.generator_jacobian(G)
-
-    traj_a = randn(Float64, 2, 3)
-    a₀ = traj_a[:, 1]
-    aᵥ = @views traj_a[:, 1]
-
-    @test ∂G(a₀) isa AbstractVector{<:AbstractMatrix{Float64}}
-    @test ∂G(a₀)[1] isa AbstractMatrix
-
-    @test ∂G(aᵥ) isa AbstractVector{<:AbstractMatrix{Float64}}
-    @test ∂G(aᵥ)[1] isa AbstractMatrix{Float64}
-end
-
 end
